CREATING REST API
MODEL: The model has two fields: the task content (as a CharField) and the Boolean type whether or not the task is completed.

VIEW: There are two main views, the first one gets the list of all the tasks and it also creates a new task depending on the
method of the request. The other one, it's used when I need to pass a pk through the id, so this is used to get the task detail,
update the task or delete it. The functionality depends on the HTTP method.

SERIALIZER: I created a TaskSerializer based on my model, so I had to use the class Meta and inside of it I placed the model
attribute and also the fields that I want the serializer to work with.