Updating App Component
The app component is the only one that I'm using for the moment. This component has state: the incoming tasks list
form the API and the text shown in the input. The useEffect hook gets the tasks from the REST API using fetch,
it parses the response as JSON and uses setTasks to update the state, so the template can render the tasks list.

The onSubmit form attribute triggers the submitTask function that creates a new object and send it to the API through
a POST request. Then, inside the Promise, it changes the state pushing the data item in the array and using setTasks.