Edit and Delete Features
EDITING: The button triggers 'handleEdit' function. This is going to set the text state with the task content, the editing state as True and
ActiveItem with the task passed through the parameters. Then, the onSubmit function checks if edit is true and sends a PUT request to the
corresponding url. Finally, this request returns an object with the updated data and the tasks list is updated with the new information.

DELETE: the handleDelete function deletes the task from the list and updates the state. Then, it makes a DELETE request and deletes the
database record.