Complete Feature
Every time the user clicks on the task item, I want the frontend to set the task attribute as incomplete or complete depending on its
previous value. Then, it updates the tasks list state in order to show the strike feature. It also makes a PUT request in order to
update the database record, sending the complete and text keys in the JSON body. The div task wrapper sets 'text-decoration' as 'line-through'
if the task is completed or 'none' if that value is false.